pdx201: audio: Disable incorrect microphone gain reduction
The microphones in some recording scenarios (generally surrounding
stereo and unprocessed or fluence contexts) are almost inaudible.

On stock this is adressed by routing to different paths that do not use
this calibration:

    SND_DEVICE_IN_SONY_MONO_MIC -> speaker-mono-mic
    SND_DEVICE_IN_SONY_STEREO_MIC -> stereo-mic
    etcetera.

While we can attempt to mimic this behaviour by overriding random
SND_DEVICE_IN devices in audio_platform_info the CAF audio paths work
just as well (and are in most cases identical) when this gain adjustment
is removed.

Change-Id: I8663ea086abc1f0c6e228c0adfa4aa309402eecf
Signed-off-by: Marijn Suijten <marijns95@gmail.com>